co-gen/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database clients
chromadb==0.4.24
psycopg2-binary==2.9.9

# Document processing
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
PyPDF2==3.0.1

# ML and embeddings
sentence-transformers==2.3.1
torch==2.2.0
numpy==1.26.3

# LLM - Updated version for compatibility
groq==0.9.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
